+++ src/graphql/netflow.rs
@@ -78,6 +78,7 @@ impl RawEventFilter for NetflowFilter {
         _text: Option<String>,
         _sensor: Option<String>,
         _agent_id: Option<String>,
+        _service_name: Option<String>,
     ) -> Result<bool> {
         if check_address(self.orig_addr.as_ref(), orig_addr)?
             && check_address(self.resp_addr.as_ref(), resp_addr)?
@@ -843,6 +844,7 @@ mod tests {
                 None,
                 None,
                 None,
+                None,
             )
             .unwrap();
         assert!(ok);
@@ -855,6 +857,7 @@ mod tests {
                 Some(3500),
                 None,
                 Some("haystack".to_string()),
+                None,
                 None,
                 None,
                 None,