+++ src/graphql/timeseries.rs
@@ -54,6 +54,7 @@ impl RawEventFilter for TimeSeriesFilter {
         _text: Option<String>,
         _sensor: Option<String>,
         _agent_id: Option<String>,
+        _service_name: Option<String>,
     ) -> Result<bool> {
         Ok(true)
     }